requirements.txt
asgiref==3.7.2
feedparser==6.0.11
requests==2.31.0
bs4==0.0.2
lxml==5.1.0
pandas==2.2.1
tabulate==0.9.0
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
packaging==23.2
newspaper3k==0.2.8
textblob==0.18.0.post0
nltk==3.8.1

